pyproject.toml
[tool.poetry]
name = "modelhub"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = { version = "^0.22.0", extras = ["standard"] }
gunicorn = "^21.2.0"
pydantic = "^2"
pydantic-settings = "^2"
yarl = "^1.9.2"
ujson = "^5.8.0"
SQLAlchemy = {version = "^2.0.18", extras = ["asyncio"]}
aiosqlite = "^0.19.0"
httptools = "^0.6.0"
fastapi-restful = "^0.5.0"
typing-inspect = "^0.9.0"
optuna = "^3.4.0"
numpy = "^1.26.2"
scikit-learn = "^1.3.2"
xgboost = "^2.0.2"
azure-core = "^1.29.5"
azure-cosmos = "^4.5.1"
python-dotenv = "^1.0.0"
pandas = "^2.1.3"
azure-storage-blob = "^12.19.0"
azure-identity = "^1.15.0"


[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~4.0.1"
mypy = "^1.1.1"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^22.12.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
httpx = "^0.23.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["modelhub",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "MODELHUB_ENVIRONMENT=pytest",
    "MODELHUB_DB_FILE=test_db.sqlite3",
]

[fastapi-template.options]
project_name = "modelhub"
api_type = "rest"
enable_redis = "None"
enable_rmq = "None"
ci_type = "gitlab_ci"
enable_migrations = "None"
enable_taskiq = "None"
enable_kube = "None"
kube_name = "modelhub"
enable_routers = "True"
enable_kafka = "None"
enable_loguru = "None"
traefik_labels = "None"
add_dummy = "True"
orm = "sqlalchemy"
self_hosted_swagger = "None"
prometheus_enabled = "None"
sentry_enabled = "None"
otlp_enabled = "None"
pydanticv1 = "None"
gunicorn = "True"
add_users = "None"
cookie_auth = "None"
jwt_auth = "None"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
